% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_guide.R
\name{create_study_guide}
\alias{create_study_guide}
\title{Create a study guide as an HTML document}
\usage{
create_study_guide(guide = NULL, file_name = NULL)
}
\arguments{
\item{guide}{A character string specifying which guide to create,
without the .qmd extension. Currently supports \code{"discrete} and
\code{"continious"}.}

\item{file_name}{A character string for file name. If null, defaults to
being named after the guide specified.}
}
\value{
Called for its side effects. Creates a .qmd file in the current
  working directory and opens it in RStudio if available.
}
\description{
Creates a Quarto study guide template in the current working directory
and opens it for editing.
}
\examples{
\dontrun{
# Create a study guide
create_homework_html(guide = "discrete", file_name = "my_study_guide")

# Create a homework file without specifying a file name
create_homework_html(guide = "discrete")
}
}
